+++ src/main/java/io/github/easyretrofit/core/RetrofitInterceptorExtension.java
@@ -2,7 +2,7 @@ package io.github.easyretrofit.core;
 
 import io.github.easyretrofit.core.exception.RetrofitExtensionException;
 import io.github.easyretrofit.core.extension.BaseInterceptor;
-import io.github.easyretrofit.core.proxy.BaseExceptionDelegate;
+import io.github.easyretrofit.core.delegate.BaseExceptionDelegate;
 
 import java.lang.annotation.Annotation;
