+++ src/main/java/io/github/easyretrofit/core/proxy/RetrofitServiceProxy.java
@@ -1,5 +1,8 @@
 package io.github.easyretrofit.core.proxy;
 
+import io.github.easyretrofit.core.delegate.BaseExceptionDelegate;
+import io.github.easyretrofit.core.delegate.ExceptionDelegator;
+import io.github.easyretrofit.core.delegate.ProxyExceptionHandler;
 import io.github.easyretrofit.core.exception.RetrofitExtensionException;
 
 import java.lang.reflect.InvocationHandler;
@@ -27,32 +30,10 @@ public class RetrofitServiceProxy<T> implements InvocationHandler {
         try {
             return method.invoke(t, args);
         } catch (Exception e) {
-            Throwable cause = e.getCause();
-            if (cause instanceof RetrofitExtensionException) {
-                return getProxyExceptionObject(proxy, method, args, cause);
-            }
-//            else {
-//                Throwable targetCause = ((InvocationTargetException) e).getTargetException();
-//                if (targetCause instanceof RetrofitExtensionException) {
-//                    return getProxyExceptionObject(proxy, method, args, targetCause);
-//                }
-//            }
-            throw cause;
+            ProxyExceptionHandler proxyExceptionHandler = new ProxyExceptionHandler(exceptionDelegates);
+            return proxyExceptionHandler.handle(proxy, method, args, e);
         }
     }
 
-    private Object getProxyExceptionObject(Object proxy, Method method, Object[] args, Throwable cause) {
-        Object exObj = null;
-        for (BaseExceptionDelegate<? extends RetrofitExtensionException> exceptionDelegate : exceptionDelegates) {
-            if (exceptionDelegate.getExceptionClass().isAssignableFrom(cause.getClass())) {
-                ExceptionDelegator<? extends Throwable> delegator = new ExceptionDelegator<>(exceptionDelegate);
-                Object invoke = delegator.invoke(proxy, method, args, (RetrofitExtensionException) cause);
-                if (invoke != null) {
-                    exObj = invoke;
-                    break;
-                }
-            }
-        }
-        return exObj;
-    }
+
 }
